move lib/tag -> lib/api/tag
