+++ lib/api/boot/index.js
@@ -66,7 +66,7 @@ app.use('/rss', require('lib/rss'))
  * Tag API Service
  */
 
-app.use('/api', require('lib/tag'))
+app.use('/api', require('lib/api/tag'))
 
 /**
  * Topic API Service
